Fixed warning

+++ MCAdditions/Categories/UIImage+MCAdditions.m
@@ -58,7 +58,7 @@
             break;
 
         default:
-            [NSException raise:NSInvalidArgumentException format:@"Unsupported content mode: %ld", contentMode];
+            [NSException raise:NSInvalidArgumentException format:@"Unsupported content mode: %ld", (long) contentMode];
     }
 
     CGSize newSize = ratio == 0 ? size : CGSizeMake(self.size.width * ratio, self.size.height * ratio);
